+++ example/minimal/lib/main.dart
@@ -159,6 +159,9 @@ final class AppDependencies implements ScopeDependencies {
   }
 
   @override
+  void unmount() {}
+
+  @override
   Future<void> dispose() async {}
 }