+++ lib/src/scope/h_scope/scope_auto_dependency/scope_dependency/scope_dependency.dart
@@ -29,6 +29,7 @@ abstract interface class ScopeDependency {
   Stream<String> init();
   Stream<String> runInit();
 
+  void unmount();
   Stream<String> dispose();
   Stream<String> runDispose();